Simplify a lot of command-testing boilerplate
