+++ test/commands/dump.unit.ts
@@ -1,42 +1,13 @@
-import chai, { expect } from "chai";
-import chaiAsPromised from "chai-as-promised";
-import { Client, Message } from "eris";
-import sinon, { SinonSandbox } from "sinon";
-import sinonChai from "sinon-chai";
-import sinonTest from "sinon-test";
-import { CommandSupport } from "../../src/Command";
+import { expect } from "chai";
+import { SinonSandbox } from "sinon";
 import command from "../../src/commands/dump";
 import { initializeCardArray } from "../../src/deck/deck";
-import { DiscordInterface } from "../../src/discord/interface";
-import { OrganiserRoleProvider } from "../../src/role/organiser";
-import { DiscordWrapperMock } from "../mocks/discord";
-import { TournamentMock } from "../mocks/tournament";
-
-chai.use(chaiAsPromised);
-chai.use(sinonChai);
-const test = sinonTest(sinon);
+import { itRejectsNonHosts, msg, support, test } from "./common";
 
 describe("command:dump", function () {
-	const support: CommandSupport = {
-		discord: new DiscordInterface(new DiscordWrapperMock()),
-		tournamentManager: new TournamentMock(),
-		organiserRole: new OrganiserRoleProvider("MC-TO")
-	};
-	const msg = new Message({ id: "007", channel_id: "foo", author: { id: "0000" } }, new Client("mock"));
 	const args = ["battlecity"];
 	before(initializeCardArray);
-	it(
-		"rejects non-hosts",
-		test(function (this: SinonSandbox) {
-			const authStub = this.stub(support.tournamentManager, "authenticateHost").rejects();
-			const listStub = this.stub(support.tournamentManager, "getConfirmed").resolves([]);
-			msg.channel.createMessage = this.spy();
-			expect(command.executor(msg, args, support)).to.be.rejected;
-			expect(authStub).to.have.been.called;
-			expect(listStub).to.not.have.been.called;
-			expect(msg.channel.createMessage).to.not.have.been.called;
-		})
-	);
+	itRejectsNonHosts(support, command, msg, ["name"]);
 	it(
 		"provides a dump of all decks",
 		test(async function (this: SinonSandbox) {
